UUUUUUUUUUUUUUUUUUUUUUUUUUUUUUUGH

+++ Content.Server/Psionics/AntiPsychicWeaponComponent.cs
@@ -54,10 +54,9 @@ namespace Content.Server.Psionics
 
         /// <summary>
         ///     How much Shock damage to take when Punish(ed) by the gods for using this weapon
-        ///     TheDen - Made this a float, dunno why it wasn't
         /// </summary>
         [DataField]
-        public float PunishSelfDamage = 20f;
+        public int PunishSelfDamage = 20;
 
         /// <summary>
         ///     How long (in seconds) should the user be stunned when punished by the gods
